restructuring and refactoring
